Refactor XRPL service files
Refactored src/services/xrpl.ts into smaller files for improved maintainability.  This change separates transaction handling, balance fetching, and utility functions into their own files.
