+++ src/services/xrpl/config.ts
@@ -1,0 +1,9 @@
+
+export const XRPL_SERVERS = [
+  "wss://xrplcluster.com",
+  "wss://s1.ripple.com",
+  "wss://s2.ripple.com",
+  "wss://rippleitin.com",
+  "wss://xrpl.ws",
+  "wss://s.altnet.rippletest.net"
+] as const;
